16 buttons for QFY board implemented

+++ esphome/components/tm1638/binary_sensor/tm1638_key.cpp
@@ -3,7 +3,7 @@
 namespace esphome {
 namespace tm1638 {
 
-void TM1638Key::keys_update(uint8_t keys) {
+void TM1638Key::keys_update(uint16_t keys) {
   bool pressed = keys & (1 << key_code_);
   if (pressed != this->state)
     this->publish_state(pressed);
